Implementation of Scrap Wheelbarrow (Shopping Cart)
Added:
- Script relative to the Shopping Cart (inherited from WheelbarrowScript)
- Script relative to the Wheelbarrow that you can buy from store (inherited from WheelbarrowScript)
- Registering of the Shopping Cart's prefab in Plugin.cs
- Respective configuration options for the Shopping Cart

Changed:
- Unity assets for both prefabs (yet to use the actual models)
