+++ MoreShipUpgrades/Misc/AssetBundleHandler.cs
@@ -50,7 +50,8 @@ namespace MoreShipUpgrades.Misc
             { "Medkit", "Assets/ShipUpgrades/MedKitItem.asset" },
             { "Night Vision", "Assets/ShipUpgrades/NightVisionItem.asset" },
             { "Diving Kit", "Assets/ShipUpgrades/DivingKitItem.asset" },
-            { "Wheelbarrow", "Assets/ShipUpgrades/Items/WheelbarrowItem.asset" },
+            { "Store Wheelbarrow", "Assets/ShipUpgrades/Items/Wheelbarrow/StoreWheelbarrowItem.asset" },
+            { "Scrap Wheelbarrow", "Assets/ShipUpgrades/Items/Wheelbarrow/ScrapWheelbarrowItem.asset" },
 
             { "Break", "Assets/ShipUpgrades/break.mp3" },
             { "Error", "Assets/ShipUpgrades/error.mp3" },
